Updated the Map Generator file
Added a bit more into the block class and developed the randomization method. It currently works a bit too well. It'll be toned down soon. Added options for 5 different map sizes. A GUI will be implemented soon so that the window is sized to fit the map that is created.